Extract ProjectPage and ProjectOverviewPage components
